requirements.txt
colorama>=0.4.6
lxml>=5.2.1
more-itertools>=10.3.0
numpy<2
onnxruntime-gpu>=1.17.1
opencv-python>=4.9.0.80
pdf2image>=1.17.0
PyYAML>=6.0.1
termcolor>=2.4.0
tqdm>=4.66.2
